Android: fix NDK build warnings

+++ crengine/include/lvstring.h
@@ -586,7 +586,7 @@ public:
     /// calculates hash for string
     lUInt32 getHash() const;
     /// returns character at specified position, with index bounds checking, fatal error if fails
-    value_type & at( size_type pos ) { if (pos>pchunk->len) crFatalError(); return modify()[pos]; }
+    value_type & at( size_type pos ) { if (pos > (unsigned)pchunk->len) crFatalError(); return modify()[pos]; }
     /// returns character at specified position, without index bounds checking
     value_type operator [] ( size_type pos ) const { return pchunk->buf16[pos]; }
     /// returns reference to specified character position (lvalue)
